pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "late"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A CLI and web UI for scheduling and running training jobs on ROCm servers."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "pyyaml",
    "flask",
    "waitress",
    "transformers>=4.40.0",
    "datasets",
    "accelerate",
    "trl",
    "peft",
    "wandb",
    "torch",
    "Jinja2"
]

[project.urls]
Homepage = "https://github.com/TesslateAI/Late"
Issues = "https://github.com/TesslateAI/Late/issues"

[project.scripts]
late = "late.cli:cli"